Add script to build and pack modules.
