+++ config/AutoRestConfig.yaml
@@ -2,9 +2,10 @@
 powershell: true
 version: latest
 use: '@microsoft.azure/autorest.powershell@beta'
-module-version: 0.0.1
+module-version: 0.0.0.1
 # prefix: Graph
 subject-prefix: ''
+azure: false
 # module-name: $(prefix).$(title)
 module-name: Graph.$(title)
 namespace: Microsoft.Graph.PowerShell
