+++ src/Authentication/Authentication/Constants.cs
@@ -1,0 +1,10 @@
+﻿// ------------------------------------------------------------------------------
+//  Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
+// ------------------------------------------------------------------------------
+namespace Microsoft.Graph.PowerShell.Authentication
+{
+    public static class Constants
+    {
+        public const string GraphAuthConfigId = "GraphAuthConfigId";
+    }
+}
