+++ src/Authentication/Authentication/Graph.Authentication.nuspec
@@ -1,0 +1,25 @@
+<?xml version="1.0"?>
+<package >
+  <metadata>
+    <version>0.0.0.1</version>
+    <id>Graph.Authentication</id>
+    <title>Microsoft Graph PowerShell Authentication Module</title>
+    <authors>Microsoft</authors>
+    <owners>Microsoft</owners>
+    <requireLicenseAcceptance>true</requireLicenseAcceptance>
+    <licenseUrl>https://aka.ms/devservicesagreement</licenseUrl>
+    <iconUrl>https://raw.githubusercontent.com/microsoftgraph/g-raph/master/g-raph.png</iconUrl>
+    <projectUrl>https://github.com/microsoftgraph/msgraph-sdk-powershell</projectUrl>
+    <description>Microsoft Graph PowerShell authentication module</description>
+    <copyright>© Microsoft Corporation. All rights reserved.</copyright>
+    <tags>Microsoft Office365 Graph PowerShell GraphServiceClient Outlook OneDrive AzureAD GraphAPI Productivity SharePoint Intune SDK</tags>
+    <releaseNotes></releaseNotes>
+  </metadata>
+  <files>
+    <file src="Graph.Authentication.psd1" />
+    <!--<file src="Graph.Authentication.psm1" />-->
+    <!-- https://github.com/NuGet/Home/issues/3584 -->
+    <file src="bin/Graph.Authentication.dll" target="bin" />
+    <file src="bin/Graph.Authentication.deps.json" target="bin" />
+  </files>
+</package>
