+++ src/Authentication/Authentication/build-module.ps1
@@ -1,0 +1,51 @@
+﻿param([switch]$Isolated, [switch]$Pack, [switch]$Release)
+$ErrorActionPreference = 'Stop'
+
+if($PSEdition -ne 'Core') {
+  Write-Error 'This script requires PowerShell Core to execute. [Note] Generated cmdlets will work in both PowerShell Core or Windows PowerShell.'
+}
+
+if(-not $Isolated) {
+  Write-Host -ForegroundColor Green 'Creating isolated process...'
+  $pwsh = [System.Diagnostics.Process]::GetCurrentProcess().Path
+  & "$pwsh" -NonInteractive -NoLogo -NoProfile -File $MyInvocation.MyCommand.Path @PSBoundParameters -Isolated
+
+  if($LastExitCode -ne 0) {
+    # Build failed. Don't attempt to run the module.
+    return
+  }
+
+  if($Pack) {
+    . (Join-Path $PSScriptRoot 'pack-module.ps1')
+    if($LastExitCode -ne 0) {
+      # Packing failed. Don't attempt to run the module.
+      return
+    }
+  }
+
+  return
+}
+
+Write-Host -ForegroundColor Green 'Cleaning build folders...'
+$binFolder = Join-Path $PSScriptRoot 'bin'
+$objFolder = Join-Path $PSScriptRoot 'obj'
+$null = Remove-Item -Recurse -ErrorAction SilentlyContinue -Path $binFolder, $objFolder
+
+if((Test-Path $binFolder) -or (Test-Path $objFolder)) {
+  Write-Host -ForegroundColor Cyan 'Did you forget to exit your isolated module session before rebuilding?'
+  Write-Error 'Unable to clean ''bin'' or ''obj'' folder. A process may have an open handle.'
+}
+
+Write-Host -ForegroundColor Green 'Compiling module...'
+$buildConfig = 'Debug'
+if($Release) {
+  $buildConfig = 'Release'
+}
+dotnet publish $PSScriptRoot --verbosity quiet --configuration $buildConfig /nologo
+
+if($LastExitCode -ne 0) {
+  Write-Error 'Compilation failed.'
+}
+
+$null = Remove-Item -Recurse -ErrorAction SilentlyContinue -Path (Join-Path $binFolder 'Debug'), (Join-Path $binFolder 'Release')
+Write-Host -ForegroundColor Green '-------------Done-------------'
