+++ src/Authentication/Authentication/pack-module.ps1
@@ -1,0 +1,3 @@
+﻿Write-Host -ForegroundColor Green 'Packing Graph module...'
+dotnet pack $PSScriptRoot --no-build /nologo
+Write-Host -ForegroundColor Green '-------------Done-------------'
