--- src/Common/Common/Art/WelcomeText.txt
@@ -1,7 +1,0 @@
-﻿#     #                                                              #     #                                                       #####                                 ######                               #####                                  #####  ######  #    # ### 
-#  #  # ###### #       ####   ####  #    # ######    #####  ####     ##   ## #  ####  #####   ####   ####   ####  ###### #####    #     # #####    ##   #####  #    #    #     #  ####  #    # ###### #####  #     # #    # ###### #      #         #     # #     # #   #  ### 
-#  #  # #      #      #    # #    # ##  ## #           #   #    #    # # # # # #    # #    # #    # #      #    # #        #      #       #    #  #  #  #    # #    #    #     # #    # #    # #      #    # #       #    # #      #      #         #       #     # #  #   ### 
-#  #  # #####  #      #      #    # # ## # #####       #   #    #    #  #  # # #      #    # #    #  ####  #    # #####    #      #  #### #    # #    # #    # ######    ######  #    # #    # #####  #    #  #####  ###### #####  #      #          #####  #     # ###     #  
-#  #  # #      #      #      #    # #    # #           #   #    #    #     # # #      #####  #    #      # #    # #        #      #     # #####  ###### #####  #    #    #       #    # # ## # #      #####        # #    # #      #      #               # #     # #  #       
-#  #  # #      #      #    # #    # #    # #           #   #    #    #     # # #    # #   #  #    # #    # #    # #        #      #     # #   #  #    # #      #    #    #       #    # ##  ## #      #   #  #     # #    # #      #      #         #     # #     # #   #  ### 
- ## ##  ###### ######  ####   ####  #    # ######      #    ####     #     # #  ####  #    #  ####   ####   ####  #        #       #####  #    # #    # #      #    #    #        ####  #    # ###### #    #  #####  #    # ###### ###### ######     #####  ######  #    # ###
