--- src/Common/Common/Constants.cs
@@ -1,10 +1,0 @@
-﻿// ------------------------------------------------------------------------------
-//  Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
-// ------------------------------------------------------------------------------
-namespace Microsoft.Graph.PowerShell.Common
-{
-    public static class Constants
-    {
-        public const string GraphAuthConfigId = "GraphAuthConfigId";
-    }
-}
