--- src/Common/Common/Graph.psm1
@@ -1,108 +1,0 @@
-﻿# # region Generated 
-#   # Load the private module dll
-#   $null = Import-Module -Name (Join-Path $PSScriptRoot '\bin\Common.dll')
-
-#   # Export nothing to clear implicit exports
-#   Export-ModuleMember
-
-#   # Export script cmdlets
-#   #Get-ChildItem -Path $PSScriptRoot -Recurse -Include '*.ps1' -File | ForEach-Object { . $_.FullName }
-#   #Export-ModuleMember -Function (Get-ScriptCmdlet -ScriptFolder $PSScriptRoot) -Alias (Get-ScriptCmdlet -ScriptFolder $PSScriptRoot -AsAlias)
-# # endregion
-
-# # Import-Module -Name (Join-Path $PSScriptRoot '..\bin\Common.dll')
-# Import-Module Az.Accounts -MinimumVersion 1.6.1 -Scope Global
-
-
-# NEW
-$PSDefaultParameterValues.Clear()
-Set-StrictMode -Version Latest
-
-function Test-DotNet
-{
-    try
-    {
-        if ((Get-PSDrive 'HKLM' -ErrorAction Ignore) -and (-not (Get-ChildItem 'HKLM:\SOFTWARE\Microsoft\NET Framework Setup\NDP\v4\Full\' -ErrorAction Stop | Get-ItemPropertyValue -ErrorAction Stop -Name Release | Where { $_ -ge 461808 })))
-        {
-            throw ".NET Framework versions lower than 4.7.2 are not supported in Az.  Please upgrade to .NET Framework 4.7.2 or higher."
-        }
-    }
-    catch [System.Management.Automation.DriveNotFoundException]
-    {
-        Write-Verbose ".NET Framework version check failed."
-    }
-}
-
-if ($true -and ($PSEdition -eq 'Desktop'))
-{
-    if ($PSVersionTable.PSVersion -lt [Version]'5.1')
-    {
-        throw "PowerShell versions lower than 5.1 are not supported in Az. Please upgrade to PowerShell 5.1 or higher."
-    }
-
-    Test-DotNet
-}
-
-if (Test-Path -Path "$PSScriptRoot\StartupScripts" -ErrorAction Ignore)
-{
-    Get-ChildItem "$PSScriptRoot\StartupScripts" -ErrorAction Stop | ForEach-Object {
-        . $_.FullName
-    }
-}
-
-if (Get-Module AzureRM.profile -ErrorAction Ignore)
-{
-    Write-Warning ("AzureRM.Profile already loaded. Az and AzureRM modules cannot be imported in the same session or used in the same script or runbook. If you are running PowerShell in an environment you control you can use the 'Uninstall-AzureRm' cmdlet to remove all AzureRm modules from your machine. " +
-        "If you are running in Azure Automation, take care that none of your runbooks import both Az and AzureRM modules. More information can be found here: https://aka.ms/azps-migration-guide.")
-    throw ("AzureRM.Profile already loaded. Az and AzureRM modules cannot be imported in the same session or used in the same script or runbook. If you are running PowerShell in an environment you control you can use the 'Uninstall-AzureRm' cmdlet to remove all AzureRm modules from your machine. " +
-        "If you are running in Azure Automation, take care that none of your runbooks import both Az and AzureRM modules. More information can be found here: https://aka.ms/azps-migration-guide.")
-}
-
-$preloadPath = (Join-Path $PSScriptRoot -ChildPath "PreloadAssemblies")
-if($PSEdition -eq 'Desktop' -and (Test-Path $preloadPath -ErrorAction Ignore))
-{
-    try
-    {
-        Get-ChildItem -ErrorAction Stop -Path $preloadPath -Filter "*.dll" | ForEach-Object {
-            Add-Type -Path $_.FullName -ErrorAction Ignore | Out-Null
-        }
-    }
-    catch {}
-}
-
-$netCorePath = (Join-Path $PSScriptRoot -ChildPath "NetCoreAssemblies")
-if($PSEdition -eq 'Core' -and (Test-Path $netCorePath -ErrorAction Ignore))
-{
-    try
-    {
-        $loadedAssemblies = ([System.AppDomain]::CurrentDomain.GetAssemblies() | %{New-Object -TypeName System.Reflection.AssemblyName -ArgumentList $_.FullName} )
-        Get-ChildItem -ErrorAction Stop -Path $netCorePath -Filter "*.dll" | ForEach-Object {
-            $assemblyName = ([System.Reflection.AssemblyName]::GetAssemblyName($_.FullName))
-            $matches = ($loadedAssemblies | Where-Object {$_.Name -eq $assemblyName.Name})
-            if (-not $matches)
-            {
-                Add-Type -Path $_.FullName -ErrorAction Ignore | Out-Null
-            }
-        }
-    }
-    catch {}
-}
-
-
-$module = Get-Module Az.Accounts 
-if ($module -ne $null -and $module.Version.ToString().CompareTo("1.6.1") -lt 0) 
-{ 
-    Write-Error "This module requires Az.Accounts version 1.6.1. An earlier version of Az.Accounts is imported in the current PowerShell session. Please open a new session before importing this module. This error could indicate that multiple incompatible versions of the Azure PowerShell cmdlets are installed on your system. Please see https://aka.ms/azps-version-error for troubleshooting information." -ErrorAction Stop 
-} 
-elseif ($module -eq $null) 
-{ 
-    Import-Module Az.Accounts -MinimumVersion 1.6.1 -Scope Global 
-}
-
-
-if (Test-Path -Path "$PSScriptRoot\PostImportScripts" -ErrorAction Ignore)
-{
-    Get-ChildItem "$PSScriptRoot\PostImportScripts" -ErrorAction Stop | ForEach-Object {
-        . $_.FullName
-    }
-}
