--- src/Common/Common/Properties/launchSettings.json
@@ -1,10 +1,0 @@
-{
-  "profiles": {
-    "Common": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "Key": "Value"
-      }
-    }
-  }
-}
