--- src/Common/Common/run-module.ps1
@@ -1,1 +1,0 @@
-﻿# coming-soon
