--- src/Common/Common/test-module.ps1
@@ -1,1 +1,0 @@
-﻿# coming-soon
