--- src/me.message/me.message/.gitattributes
@@ -1,1 +1,0 @@
-* text=auto
