--- src/me.message/me.message/.gitignore
@@ -1,14 +1,0 @@
-bin
-obj
-.vs
-generated
-internal
-exports
-custom/*.psm1
-test/*-TestResults.xml
-/*.ps1
-/*.ps1xml
-/*.psm1
-/*.snk
-/*.csproj
-/*.nuspec
