--- src/me.message/me.message/how-to.md
@@ -1,48 +1,0 @@
-# How-To
-This document describes how to develop for `Graph.me.message`.
-
-## Building `Graph.me.message`
-To build, run the `build-module.ps1` at the root of the module directory. This will generate the proxy script cmdlets that are the cmdlets being exported by this module. After the build completes, the proxy script cmdlets will be output to the `exports` folder. To read more about the proxy script cmdlets, look at the [readme.md](exports/readme.md) in the `exports` folder.
-
-## Creating custom cmdlets
-To add cmdlets that were not generated by the REST specification, use the `custom` folder. This folder allows you to add handwritten `.ps1` and `.cs` files. Currently, we support using `.ps1` scripts as new cmdlets or as additional low-level variants (via `ParameterSet`), and `.cs` files as low-level (variants) cmdlets that the exported script cmdlets call. We do not support exporting any `.cs` (dll) cmdlets directly. To read more about custom cmdlets, look at the [readme.md](custom/readme.md) in the `custom` folder.
-
-## Generating documentation
-To generate documentation, run the `generate-help.ps1` script. This will look at the documentation comments in the generated and custom cmdlets and types, and create `.md` files into the `docs` folder. Additionally, this pulls in any examples from the `examples` folder and adds them to the generated help markdown documents. To read more about examples, look at the [readme.md](examples/readme.md) in the `examples` folder. To read more about documentation, look at the [readme.md](docs/readme.md) in the `docs` folder.
-
-## Testing `Graph.me.message`
-To test the cmdlets, we use [Pester](https://github.com/pester/Pester). Tests scripts (`.ps1`) should be added to the `test` folder. To execute the Pester tests, run the `test-module.ps1` script. This will run all tests in `playback` mode within the `test` folder. To read more about testing cmdlets, look at the [readme.md](examples/readme.md) in the `examples` folder.
-
-## Packing `Graph.me.message`
-To pack `Graph.me.message` for distribution, run the `pack-module.ps1` script. This will take the contents of multiple directories and certain root-folder files to create a `.nupkg`. The structure of the `.nupkg` is created so it can be loaded part of a [PSRepository](https://docs.microsoft.com/en-us/powershell/module/powershellget/register-psrepository). Additionally, this package is in a format for distribution to the [PSGallery](https://www.powershellgallery.com/). For signing an Azure module, please contact the [Azure PowerShell](https://github.com/Azure/azure-powershell) team.
-
-## Module Script Details
-There are multiple scripts created for performing different actions for developing `Graph.me.message`.
-- `build-module.ps1`
-  - Builds the module DLL (`./bin/Graph.me.message.private.dll`), creates the exported cmdlets, generates custom cmdlet test stubs, and updates `./Graph.me.message.psd1` with Azure profile information.
-  - **Parameters**: [`Switch` parameters]
-    - `-Run`: After building, creates an isolated PowerShell session and loads `Graph.me.message`.
-    - `-Test`: After building, runs the `Pester` tests defined in the `test` folder.
-    - `-Docs`: After building, generates the Markdown documents for the modules into the `docs` folder.
-    - `-Pack`: After building, packages the module into a `.nupkg`.
-    - `-Code`: After building, opens a VSCode window with the module's directory and runs (see `-Run`) the module.
-    - `-Release`: Builds the module in `Release` configuration (as opposed to `Debug` configuration).
-- `run-module.ps1`
-  - Creates an isolated PowerShell session and loads `Graph.me.message` into the session.
-  - Same as `-Run` in `build-module.ps1`.
-  - **Parameters**: [`Switch` parameters]
-    - `-Code`: Opens a VSCode window with the module's directory.
-      - Same as `-Code` in `build-module.ps1`.
-- `generate-help.ps1`
-  - Generates the Markdown documents for the modules into the `docs` folder.
-  - Same as `-Docs` in `build-module.ps1`.
-- `test-module.ps1`
-  - Runs the `Pester` tests defined in the `test` folder.
-  - Same as `-Test` in `build-module.ps1`.
-- `pack-module.ps1`
-  - Packages the module into a `.nupkg` for distribution.
-  - Same as `-Pack` in `build-module.ps1`.
-- `check-dependencies.ps1`
-  - Used in `run-module.ps1`, `generate-help.ps1`, and `test-module.ps1` to verify dependent modules are available to run those tasks.
-  - It will download local (within the module's directory structure) versions of those modules as needed.
-  - This script *does not* need to be ran by-hand.
