--- src/me.message/me.message/readme.md
@@ -1,21 +1,0 @@
-<!-- region Generated -->
-# Graph.me.message
-This directory contains the PowerShell module for the MeMessage service.
-
----
-## Status
-[![Graph.me.message](https://img.shields.io/powershellgallery/v/Graph.me.message.svg?style=flat-square&label=Graph.me.message "Graph.me.message")](https://www.powershellgallery.com/packages/Graph.me.message/)
-
-## Info
-- Modifiable: yes
-- Generated: all
-- Committed: yes
-- Packaged: yes
-
----
-## Detail
-This module was primarily generated via [AutoRest](https://github.com/Azure/autorest) using the [PowerShell](https://github.com/Azure/autorest.powershell) extension.
-
-## Development
-For information on how to develop for `Graph.me.message`, see [how-to.md](how-to.md).
-<!-- endregion -->
