--- src/me.message/me.message/resources/readme.md
@@ -1,11 +1,0 @@
-# Resources
-This directory can contain any additional resources for module that are not required at runtime. This directory **does not** get packaged with the module. If you have assets for custom implementation, place them into the `..\custom` folder.
-
-## Info
-- Modifiable: yes
-- Generated: no
-- Committed: yes
-- Packaged: no
-
-## Purpose
-Use this folder to put anything you want to keep around as part of the repository for the module, but is not something that is required for the module. For example, development files, packaged builds, or additional information. This is only intended to be used in repositories where the module's output directory is cleaned, but tangential resources for the module want to remain intact.
