--- src/me.message/me.message/test/readme.md
@@ -1,17 +1,0 @@
-# Test
-This directory contains the [Pester](https://www.powershellgallery.com/packages/Pester) tests to run for the module. We use Pester as it is the unofficial standard for PowerShell unit testing. Test stubs for custom cmdlets (created in `..\custom`) will be generated into this folder when `build-module.ps1` is ran. These test stubs will fail automatically, to indicate that tests should be written for custom cmdlets.
-
-## Info
-- Modifiable: yes
-- Generated: partial
-- Committed: yes
-- Packaged: no
-
-## Details (*WIP*)
-We allow three testing modes: `live`, `record`, and `playback`.
-
-## Purpose
-Custom cmdlets generally encompass additional functionality not described in the REST specification, or combines functionality generated from the REST spec. To validate this functionality continues to operate as intended, creating tests that can be ran and re-ran against custom cmdlets is part of the framework.
-
-## Usage (*WIP*)
-This feature is currently a **work-in-progress**. It is able to create test recordings of the HTTP pipeline. However, folder structure, file names, and processes are being implemented.
