+++ tools/BuildAndPackModule.ps1
@@ -1,0 +1,65 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+<#
+.DESCRIPTION
+Builds and packs generated module into nuget packages for distribution. The generated packages are moved to '{repo}\bin' by default for easy access.
+
+.PARAMETER Module
+The name of the module to build and pack.
+
+.PARAMETER OutputFolder
+The output folder to move the generated nuget packages to. This defaults to '{repo}\artifacts'.
+#>
+Param(
+    [Parameter(Mandatory = $true)]
+    [ValidateNotNullOrEmpty()]
+    [string] $Module,
+    [string] $OutputFolder,
+    [string[]] $RequiredModules
+)
+
+$ModuleDir = "./src/$Module/$Module"
+$BuildModulePS1 = Join-Path $moduleDir "/build-module.ps1"
+$PackModulePS1 = Join-Path $moduleDir "/pack-module.ps1"
+
+if([string]::IsNullOrEmpty($OutputFolder)) {
+    $OutputFolder = "./artifacts/"
+}
+
+if (-not (Test-Path -Path $BuildModulePS1)){
+    Write-Host -ForegroundColor Red "Build script file '$BuildModulePS1' not found for $Module module."
+}
+
+# Build module
+& $BuildModulePS1 -Test -Docs -Release
+if($LastExitCode -ne 0) {
+    # Build failed, don't pack the module.
+    Write-Error "Failed to build $Module."
+    return
+}
+
+if($RequiredModules.Count -gt 0) {
+    # Add required modules.
+    Update-ModuleManifest -Path (Join-Path $ModuleDir "Graph.$Module.psd1") -RequiredModules $RequiredModules
+}
+
+# Pack module
+& $PackModulePS1
+
+if($LastExitCode -ne 0) {
+    # Pack failed, don't attempt to move nuget package.
+    Write-Error "Failed to pack $Module."
+    return
+}
+
+# Get generated .nupkg
+$NugetPackage = (Get-ChildItem (Join-Path $ModuleDir "./bin") | Where-Object Name -Match ".nupkg").FullName
+
+if(-not (Test-Path $OutputFolder)) {
+    # Create artifacts folder.
+    New-Item -Path $OutputFolder -Type Directory
+}
+
+# Copy package to artifacts folder.
+Move-Item -Path $NugetPackage -Destination $OutputFolder -Force
+Write-Host -ForegroundColor Green "Moved '$NugetPackage' to '$OutputFolder'."
