+++ tools/GenerateRollupModule.ps1
@@ -1,0 +1,10 @@
+# Generate Graph module
+# Roll-up generated modules
+
+#1. Get successfully generated modules.
+
+#2. New-ModuleManifest specifying generated modules as dependencies.
+
+#3. New-Module that loads manifest's dependencies.
+
+#4. Pack module into .nupkg.
